changed ld, st, zjmp, fixed set_mem_value segv

+++ corewar/src/instructions/sti.c
@@ -6,7 +6,7 @@
 /*   By: vsanchez <vsanchez@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/26 05:10:09 by vsanchez          #+#    #+#             */
-/*   Updated: 2019/08/08 14:20:28 by mbeilles         ###   ########.fr       */
+/*   Updated: 2019/08/09 14:23:48 by vsanchez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,6 +33,6 @@ void				op_sti(t_process *proc)
 	else if (proc->op.types[2] == COR_ARG_DIR)
 		value_2 = (uint16_t)proc->op.args[2];
 	set_mem_value(proc->global_offset + proc->pc
-			+ (value_1 + value_2) % COR_IDX_MOD,
+			+ (int16_t)(value_1 + value_2) % COR_IDX_MOD,
 			get_reg_value(proc, proc->op.args[0]), COR_REG_SIZE);
 }
